+++ docs/changes.rst
@@ -9,10 +9,13 @@ To be released.
 - Now :class:`~wikidata.client.Client` became able to customize how it decodes
   datavalues to Python objects.
 
-  - Added :mod:`wikidata.datavalue` and :class:`~wikidata.datavalue.Decoder`
-    class inside it.
+  - Added :mod:`wikidata.datavalue` module and
+    :class:`~wikidata.datavalue.Decoder` class inside it.
   - Added :attr:`~.wikidata.client.Client.datavalue_decoder` option to
     :class:`~wikidata.client.Client`.
+
+- Added :mod:`wikidata.commonsmedia` module and
+  :class:`~wikidata.commonsmedia.File` class inside it.
 
 - The meaning of :class:`~wikidata.client.Client` constructor's ``base_url``
   prameter beccame not to contain the trailing path ``wiki/`` from